# swagger.yaml
openapi: 3.0.0
info:
  title: Rooms API
  description: API para gerenciamento de salas e mobílias com autenticação OAuth
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com

servers:
  - url: http://localhost:8188/api/v1
    description: Servidor de desenvolvimento - Rooms API
  - url: http://localhost:8180
    description: Servidor de desenvolvimento - OAuth

tags:
  - name: Auth
    description: Endpoints de autenticação OAuth
  - name: Rooms
    description: Gerenciamento de salas
  - name: Furnitures
    description: Gerenciamento de mobílias

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtido via endpoint de login
  
  schemas:
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: admin@pucrs.br
        password:
          type: string
          format: password
          example: a12345678
    
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
          description: Token de acesso JWT
        refresh_token:
          type: string
          description: Token de refresh
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          example: 3600
    
    Room:
      type: object
      required:
        - capacity
        - number
        - building
        - category
        - floor
        - status
      properties:
        id:
          type: string
          format: uuid
          description: ID único da sala
        capacity:
          type: integer
          minimum: 1
          example: 30
          description: Capacidade da sala
        number:
          type: string
          example: "101"
          description: Número da sala
        building:
          type: string
          example: Main
          description: Prédio onde a sala está localizada
        category:
          type: string
          enum: [CLASSROOM, LAB, OFFICE, MEETING_ROOM, AUDITORIUM]
          example: CLASSROOM
          description: Categoria da sala
        floor:
          type: integer
          example: 1
          description: Andar da sala
        description:
          type: string
          example: Room for data engineering classes
          description: Descrição da sala
        status:
          type: string
          enum: [ACTIVE, INACTIVE, MAINTENANCE]
          example: ACTIVE
          description: Status da sala
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    
    CreateRoomDto:
      type: object
      required:
        - capacity
        - number
        - building
        - category
        - floor
        - status
      properties:
        capacity:
          type: integer
          minimum: 1
          example: 30
        number:
          type: string
          example: "101"
        building:
          type: string
          example: Main
        category:
          type: string
          enum: [CLASSROOM, LAB, OFFICE, MEETING_ROOM, AUDITORIUM]
          example: CLASSROOM
        floor:
          type: integer
          example: 1
        description:
          type: string
          example: Room for data engineering classes
        status:
          type: string
          enum: [ACTIVE, INACTIVE, MAINTENANCE]
          example: ACTIVE
    
    UpdateRoomDto:
      type: object
      properties:
        capacity:
          type: integer
          minimum: 1
          example: 35
        number:
          type: string
          example: "101"
        building:
          type: string
          example: Main
        category:
          type: string
          enum: [CLASSROOM, LAB, OFFICE, MEETING_ROOM, AUDITORIUM]
          example: CLASSROOM
        floor:
          type: integer
          example: 1
        description:
          type: string
          example: Updated room description
        status:
          type: string
          enum: [ACTIVE, INACTIVE, MAINTENANCE]
          example: ACTIVE
    
    PatchRoomDto:
      type: object
      properties:
        capacity:
          type: integer
          minimum: 1
          example: 40
        number:
          type: string
        building:
          type: string
        category:
          type: string
          enum: [CLASSROOM, LAB, OFFICE, MEETING_ROOM, AUDITORIUM]
        floor:
          type: integer
        description:
          type: string
        status:
          type: string
          enum: [ACTIVE, INACTIVE, MAINTENANCE]
    
    Furniture:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Table
        quantity:
          type: integer
          minimum: 1
          example: 10
        roomId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    
    CreateFurnitureDto:
      type: object
      required:
        - name
        - quantity
      properties:
        name:
          type: string
          example: Table
        quantity:
          type: integer
          minimum: 1
          example: 10
    
    Error:
      type: object
      properties:
        statusCode:
          type: integer
        message:
          type: string
        error:
          type: string

paths:
  /login:
    post:
      tags:
        - Auth
      summary: Login do usuário
      description: Autentica o usuário e retorna tokens de acesso
      operationId: login
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  example: admin@pucrs.br
                password:
                  type: string
                  example: a12345678
      responses:
        '200':
          description: Login realizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Credenciais inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  /validate:
    post:
      tags:
        - Auth
      summary: Valida o token de acesso
      description: Verifica se o token JWT é válido
      operationId: validate
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Token válido
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                    example: true
        '401':
          description: Token inválido ou expirado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  /rooms:
    get:
      tags:
        - Rooms
      summary: Lista todas as salas
      description: Retorna uma lista de todas as salas cadastradas
      operationId: listRooms
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de salas retornada com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Room'
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      tags:
        - Rooms
      summary: Cria uma nova sala
      description: Cria uma nova sala com os dados fornecidos
      operationId: createRoom
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoomDto'
      responses:
        '201':
          description: Sala criada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  /rooms/{roomId}:
    parameters:
      - name: roomId
        in: path
        required: true
        description: ID da sala
        schema:
          type: string
          format: uuid
    
    get:
      tags:
        - Rooms
      summary: Busca sala por ID
      description: Retorna os detalhes de uma sala específica
      operationId: getRoomById
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Sala encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        '404':
          description: Sala não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    put:
      tags:
        - Rooms
      summary: Atualiza sala completamente
      description: Atualiza todos os dados de uma sala existente
      operationId: updateRoom
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoomDto'
      responses:
        '200':
          description: Sala atualizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Sala não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    patch:
      tags:
        - Rooms
      summary: Atualiza sala parcialmente
      description: Atualiza apenas os campos fornecidos de uma sala
      operationId: patchRoom
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchRoomDto'
      responses:
        '200':
          description: Sala atualizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Sala não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    delete:
      tags:
        - Rooms
      summary: Remove uma sala
      description: Remove uma sala do sistema
      operationId: deleteRoom
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Sala removida com sucesso
        '404':
          description: Sala não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  /rooms/{roomId}/furnitures:
    parameters:
      - name: roomId
        in: path
        required: true
        description: ID da sala
        schema:
          type: string
          format: uuid
    
    get:
      tags:
        - Furnitures
      summary: Lista mobílias da sala
      description: Retorna todas as mobílias de uma sala específica
      operationId: listFurnitures
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de mobílias retornada com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Furniture'
        '404':
          description: Sala não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      tags:
        - Furnitures
      summary: Adiciona mobília à sala
      description: Cria uma nova mobília em uma sala específica
      operationId: createFurniture
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFurnitureDto'
      responses:
        '201':
          description: Mobília criada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Furniture'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Sala não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  /rooms/{roomId}/furnitures/{furnitureId}:
    parameters:
      - name: roomId
        in: path
        required: true
        description: ID da sala
        schema:
          type: string
          format: uuid
      - name: furnitureId
        in: path
        required: true
        description: ID da mobília
        schema:
          type: string
          format: uuid
    
    delete:
      tags:
        - Furnitures
      summary: Remove mobília
      description: Remove uma mobília específica de uma sala
      operationId: deleteFurniture
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Mobília removida com sucesso
        '404':
          description: Sala ou mobília não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'